getting a foreing key constranint on questionAttempts insert no biggie as most would have gone thru it looks like the x answers...

+++ src/breslin/serverside/battle/battle.h
@@ -28,8 +28,8 @@ Combatant* mAwayCombatant;
 
 // battle time
 long mBattleTime;
-double mBattleStartTime;
-double mBattleEndTime;
+long mBattleStartTime;
+long mBattleEndTime;
 long mBattleLengthTime;
